🏫 Feat: 기초 탭바 설계
네비게이션 뒤로가기 메모리 오류 수정
