+++ ATeen/ATeen/Scenes/MainTab/SettingsCoordinator+SettingsViewControllerCoordinator.swift
@@ -22,23 +22,17 @@ extension SettingsCoordinator: SettingsViewControllerCoordinator {
     }
     
     private func cellUserConfigurationCoordinator() {
-        userConfigurationCoordinator = factory.makeUserConfigurationCoordinator(delegate: self)
-        userConfigurationCoordinator?.start()
-        guard let userConfigurationCoordinator = userConfigurationCoordinator else { return }
+        let userConfigurationCoordinator = factory.makeUserConfigurationCoordinator(delegate: self)
+        addChildCoordinatorStart(userConfigurationCoordinator)
+      
         navigation.present(
             userConfigurationCoordinator.navigation.rootViewController,
             animated: true)
         
         userConfigurationCoordinator.navigation.dismissNavigation = { [weak self] in
-            self?.userConfigurationCoordinator = nil
+            self?.removeChildCoordinator(userConfigurationCoordinator)
         }
         
     }
 }
 
-extension SettingsCoordinator: UserConfigurationCoordinatorDelegate {
-    func didFinish() {
-        userConfigurationCoordinator = nil
-        navigation.dismiss(animated: true)
-    }
-}
