+++ ATeen/ATeen/Scenes/PostDetail/Presentation/PostDetailViewController.swift
@@ -95,5 +95,3 @@ final class PostDetailViewController: UIViewController {
     
 }
 
-// MARK: - Extensions here
-
